chore: abstract gradient stuff
