+++ src/app/page.tsx
@@ -1,71 +1,14 @@
 "use client";
 
-import { useEffect, useRef } from "react";
 import Main from "./components/Main";
 import Footer from "./components/Footer";
 import Header from "./components/Header";
+import Gradients from "./components/common/Gradients";
 
 export default function Home() {
-  const interactiveRef = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    const interBubble = interactiveRef.current;
-    if (!interBubble) return;
-
-    let curX = 0;
-    let curY = 0;
-    let tgX = 0;
-    let tgY = 0;
-
-    const move = () => {
-      curX += (tgX - curX) / 20;
-      curY += (tgY - curY) / 20;
-      interBubble.style.transform = `translate(${Math.round(
-        curX
-      )}px, ${Math.round(curY)}px)`;
-      requestAnimationFrame(move);
-    };
-
-    const handleMouseMove = (event: MouseEvent) => {
-      tgX = event.clientX;
-      tgY = event.clientY;
-    };
-
-    window.addEventListener("mousemove", handleMouseMove);
-    move();
-
-    return () => {
-      window.removeEventListener("mousemove", handleMouseMove);
-    };
-  }, []);
-
   return (
     <div className="gradient-bg">
-      <svg className="goo" xmlns="http://www.w3.org/2000/svg">
-        <defs>
-          <filter id="goo">
-            <feGaussianBlur
-              in="SourceGraphic"
-              stdDeviation="10"
-              result="blur"
-            />
-            <feColorMatrix
-              in="blur"
-              mode="matrix"
-              values="1 0 0 0 0  0 1 0 0 0  0 0 1 0 0  0 0 0 18 -8"
-              result="goo"
-            />
-            <feBlend in="SourceGraphic" in2="goo" />
-          </filter>
-        </defs>
-      </svg>
-      <div className="gradients-container">
-        <div className="g1" />
-        <div className="g2" />
-        <div className="g3" />
-        <div className="g4" />
-        <div className="g5" />
-        <div className="interactive" ref={interactiveRef} />
-      </div>
+      <Gradients />
       <div className="relative z-2 bg-black/60">
         <div className="grid grid-rows-[100px_1fr_50px] bg-black-100 min-h-screen mx-auto max-w-[1000px] gap-8 sm:p-8">
           <Header />
